+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/suggestions/commands/CommandSuggestionList.kt
@@ -52,7 +52,7 @@ public fun CommandSuggestionList(
     modifier: Modifier = Modifier,
     onCommandSelected: (Command) -> Unit = {},
     itemContent: @Composable (Command) -> Unit = { command ->
-        DefaultCommandSuggestionItem(
+        ChatTheme.componentFactory.MessageComposerCommandSuggestionItem(
             command = command,
             onCommandSelected = onCommandSelected,
         )
@@ -98,18 +98,4 @@ public fun CommandSuggestionList(
             }
         }
     }
-}
-
-/**
- * The default command suggestion item.
- *
- * @param command The given command.
- * @param onCommandSelected Handler when the command is selected.
- */
-@Composable
-internal fun DefaultCommandSuggestionItem(
-    command: Command,
-    onCommandSelected: (Command) -> Unit,
-) {
-    CommandSuggestionItem(command = command, onCommandSelected = onCommandSelected)
 }